+++ extensions/stable-api-testing/AnimalControllerServices.cpp
@@ -15,21 +15,16 @@
  * limitations under the License.
  */
 
-#pragma once
+#include "AnimalControllerServices.h"
 
-#include <cinttypes>
+#include "api/utils/ProcessorConfigUtils.h"
+namespace org::apache::nifi::minifi::api_testing {
 
-namespace org::apache::nifi::minifi::api_sandbox {
-class NumberOfLegsControllerApi {
- public:
-  virtual ~NumberOfLegsControllerApi() = default;
-  virtual uint8_t numberOfLegs() const = 0;
-};
+MinifiStatus DogController::enableImpl(api::core::ControllerServiceContext& ctx) {
+  this->has_jetpack_ = ctx.getProperty(HasJetpack) | minifi::utils::andThen(parsing::parseBool) |
+      minifi::utils::orThrow(fmt::format("Expected parsable bool from \"{}\"", HasJetpack.name));
 
-class CanFlyControllerApi {
- public:
-  virtual ~CanFlyControllerApi() = default;
-  virtual bool canFly() const = 0;
-};
+  return MINIFI_STATUS_SUCCESS;
+}
 
-}  // namespace org::apache::nifi::minifi::api_sandbox
+}  // namespace org::apache::nifi::minifi::api_testing
